Seems OK for RF.
Former-commit-id: 649b1d439e60188a82bb54cea4c0ef405bd08832

+++ sources/solvers/GREEDY/src/Explainer.h
@@ -21,7 +21,7 @@ namespace PyLE {
         int n_iterations;
         int time_limit; //in seconds
         int try_to_remove;
-        std::vector<double> excluded_features;
+        std::vector<int> excluded_features;
         Propagator::Propagator *theory_propagator = nullptr;
 
         Explainer(int _n_classes, Type t) : n_classes(_n_classes), _type(t), n_iterations(50), time_limit(0) {}
